STT: CPU fallback uses float32 (CUDA-built ctranslate2 drops int8 CPU)
The from-source CUDA build of ctranslate2 only ships float32 for CPU
in get_supported_compute_types(). int8 fallback would fail; use
float32 so the path stays viable even though it's slow.

+++ jetson/ros2_ws/src/robot_perception/robot_perception/speech_recognizer_node.py
@@ -146,8 +146,9 @@ class SpeechRecognizer(Node):
                     f'whisper load failed on device={device}: {e}. '
                     'Falling back to CPU int8.')
                 try:
+                    # CUDA-built ctranslate2 only ships float32 for CPU.
                     self._model = WhisperModel(
-                        model_size, device='cpu', compute_type='int8')
+                        model_size, device='cpu', compute_type='float32')
                     self.get_logger().info(
                         f'whisper loaded (CPU fallback): model={model_size}')
                 except Exception as e2:
